Move all attribute related code into cri_attributes module
