+++ src/cri_attributes.rs
@@ -1,9 +1,47 @@
 use std::collections::HashMap;
 
+use der_parser::ber::{ber_read_element_header, parse_ber_sequence_defined_g, BerTag};
+use der_parser::der::{der_read_element_header, parse_der_oid};
+use der_parser::error::BerError;
 use der_parser::oid::Oid;
+use nom::combinator::map_res;
+use nom::Err;
 use oid_registry::*;
 
-use crate::extensions::X509Extension;
+use crate::{
+    error::{X509Error, X509Result},
+    extensions::X509Extension,
+};
+
+/// Attributes for Certification Request
+#[derive(Debug, PartialEq)]
+pub struct X509CriAttribute<'a> {
+    pub oid: Oid<'a>,
+    pub value: &'a [u8],
+    pub(crate) parsed_attribute: ParsedCriAttribute<'a>,
+}
+
+impl<'a> X509CriAttribute<'a> {
+    pub fn from_der(i: &'a [u8]) -> X509Result<X509CriAttribute> {
+        parse_ber_sequence_defined_g(|_, i| {
+            let (i, oid) = map_res(parse_der_oid, |x| x.as_oid_val())(i)?;
+            let value_start = i;
+            let (i, hdr) = der_read_element_header(i)?;
+            if hdr.tag != BerTag::Set {
+                return Err(Err::Error(BerError::BerTypeError));
+            };
+
+            let (i, parsed_attribute) = crate::cri_attributes::parser::parse_attribute(i, &oid)?;
+            let ext = X509CriAttribute {
+                oid,
+                value: &value_start[..value_start.len() - i.len()],
+                parsed_attribute,
+            };
+            Ok((i, ext))
+        })(i)
+        .map_err(|_| X509Error::InvalidAttributes.into())
+    }
+}
 
 /// Section 3.1 of rfc 5272
 #[derive(Debug, PartialEq)]
@@ -69,4 +107,33 @@ pub(crate) mod parser {
             })
             .map_err(|_| Err::Error(BerError::BerTypeError))
     }
+}
+
+fn attributes_sequence_to_map<'a>(
+    i: &'a [u8],
+    v: Vec<X509CriAttribute<'a>>,
+) -> X509Result<'a, HashMap<Oid<'a>, X509CriAttribute<'a>>> {
+    let mut attributes = HashMap::new();
+    for attr in v.into_iter() {
+        if attributes.insert(attr.oid.clone(), attr).is_some() {
+            // duplicate attributes are not allowed
+            return Err(Err::Failure(X509Error::DuplicateAttributes));
+        }
+    }
+    Ok((i, attributes))
+}
+
+pub(crate) fn parse_cri_attributes(i: &[u8]) -> X509Result<HashMap<Oid, X509CriAttribute>> {
+    let (i, hdr) = ber_read_element_header(i).or(Err(Err::Error(X509Error::InvalidAttributes)))?;
+    if i.is_empty() {
+        return Ok((i, HashMap::new()));
+    }
+    (0..hdr.structured)
+        .into_iter()
+        .try_fold((i, Vec::new()), |(i, mut attrs), _| {
+            let (rem, attr) = X509CriAttribute::from_der(i)?;
+            attrs.push(attr);
+            Ok((rem, attrs))
+        })
+        .and_then(|(i, attrs)| attributes_sequence_to_map(i, attrs))
 }
